+++ src/App/App.js
@@ -1,6 +1,8 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
+import "@fontsource/roboto-condensed"
 
 import './App.css';
+import "../variables.css"
 
 import UI from '../Components/UI/UI';