+++ src/Components/CustomNavbar/CustomNavbar.tsx
@@ -8,7 +8,7 @@ export default function CustomNavbar() {
     return (
         <Navbar data-bs-theme="dark" className='custom-navbar'>
             <Navbar.Brand>
-                <img id="logo" src='slo.svg' alt='Logo' />
+                <img id="logo" src='res/logo.svg' alt='Logo' />
             </Navbar.Brand>
 
             <Navbar.Brand id='title'>Občinko</Navbar.Brand>