+++ src/Components/LoseScreen/LoseScreen.tsx
@@ -4,7 +4,7 @@ import { Modal } from 'react-bootstrap';
 
 export default function LoseScreen({ obcina }: {obcina: string}) {
     return (
-        <Modal className='lose-modal' show > {/* Centered?? */}
+        <Modal className='lose-modal' show centered >
             <Modal.Header>
                 <Modal.Title>Pravilna občina: { obcina } </Modal.Title>
             </Modal.Header>