+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 2025-01-21
+## v0.5 2025-01-21
 
 ### Added 
 
@@ -16,14 +16,14 @@
 - Combined win/lose game retry code to function
 
 
-## 2025-01-22
+## v0.5.1 2025-01-22
 
 ### Added
 
 - **Changelog**
 
 
-## 2025-01-24
+## v0.5.4 2025-01-24
 
 ### Added
 
@@ -36,7 +36,7 @@
 - **Converted to typescript**
 
 
-## 2025-1-26
+## v0.6 2025-1-26
 
 ### Added
 
@@ -48,13 +48,13 @@
 - Changed button
 
 
-## 2025-1-26
+## v0.6.2 2025-1-26
 
 ### Other
 
 - Fixed dropdown
 
-## 2025-1-27
+## v0.6.4 2025-1-27
 
 ### Added
 
@@ -65,13 +65,13 @@
 - Made obcine rendered(input.tsx) once
 
 
-## 2025-1-28
+## v0.6.5 2025-1-28
 
 ### Other
 
 - Improved component rendering (divs)
 
-## 2025-21-30
+## v0.7 2025-1-30
 
 ### Added 
 
@@ -83,7 +83,7 @@
 - Passing obcine to input
 
 
-## 2025-21-31
+## v0.9 2025-1-31
 
 ### Added
 
@@ -98,3 +98,26 @@
 - Region text/location
 - Dropdown size
 - Changed adjacent obcine color
+
+
+## v0.9.8 2025-2-1
+
+### Added
+
+- Css variables
+- Mobile responsiveness
+- Input validation, min/max char length
+- Zoom controls
+
+### Other 
+
+- Improved win/lose modals
+- Res folder to public + obcine
+- New tile layer with no label
+
+
+## v1.0 2025-2-2
+
+### TODO
+
+- Map zoom !